feat: beta6M10-1 grille qui s'adapte au nombre d'étapes.
